初步完成内容可以从导航栏下边穿过的效果

+++ lib/src/main/java/com/droidworker/lib/ILoadingLayout.java
@@ -9,5 +9,5 @@ public interface ILoadingLayout {
 
     int getSize();
 
-    void onPull(State state, int distance);
+    void onPull(State state, float distance);
 }
